requirements.txt
aiosmtpd==1.4.6
annotated-types==0.7.0
asgiref==3.8.1
atpublic==5.1
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-bootstrap-datepicker-plus==5.0.5
django-recaptcha==4.0.0
djangorestframework==3.16.0
drf-redesign==0.5.0
idna==3.10
inflection==0.5.1
packaging==24.2
pydantic==2.10.6
pydantic_core==2.27.2
pysimplemodel==2.3.3
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
Unipath==1.1
uritemplate==4.1.1
urllib3==2.3.0
